all, physics-bh: misc updates, more barriers, adjustments
Still working on the Barnes-Hut octree generation.

+++ resources/shaders/text_fs.glsl
@@ -1,11 +1,11 @@
 #version 100
 
-precision lowp float;
+precision highp float;
 
 varying vec2 texcoord;
 uniform sampler2D tex;
 uniform vec4 textcolor;
 
 void main(void) {
-	gl_FragColor = vec4(1, 1, 1, texture2D(tex, texcoord).a) * textcolor;
+	gl_FragColor = vec4(1, 1, 1, texture2D(tex, texcoord).a)*textcolor;
 }
